Re-write the backend API entrypoints
The MR needed to have a proper offset,
so this changes it to /mr-api. As part of this
I wanted to revisit how the MR structured
the APIs internally, preferably to match
how they are structured in the client dart code.

This makes them considerably easier to find. This
has led to a few discoveries around inconsistent data,
and minor issues with security and model structures.