Containerised Build + All in one options
This commit contains the resolution of the full
containerised build of the front end and back-ends
from a Maven build.

The Admin-UI build shells out
and runs a docker container to build the Admin-UI.

There is a new tile for managing the party server
and the MR server because they both need to contain
the admin UI.

The admin-ui builds and installs as a zip image
and while building the other docker images grab
that zip image from the repository and build it in.

This image also contains the all-in-one h2 and all-in-one
postgres (postgres is external so technically not all
in one).