Ensure promotion order for each new environment
An application once created is given a new
production environment. All subsequent
environments should insert themselves before
the first one.

This adds that functionality, it adds e2e testing
to ensure all that works.

The code that shows the user things were not
yet ordered has been removed as it is no longer
necessary.

During this phase it
was discovered that some of the tree graph
requests to get data were not operational, they
would return deleted data. This was fixed in a number
of places.

All e2e and unit and integration tests have been
run to ensure everything is still operational.

fixes issue #3