Explicitly deal with empty environments or service accounts (#9)
The Dacha Server was not able to determine that
there were no service accounts or environments, which
is typical of a new server. This would lead to
it continually trying to become master and timing
out.

This change includes the server publishing explicity
that the service accounts or environments are empty
so the cache will settle.

This has been tested with multiple Dacha servers and
they continue to negotiate between each other and
recognize the empty cache.