Conflicting users being added in Groups causing explosions
There is a bug in the ChipsInput widget that is causing
deleted items to come back as if they were added. That will
be addressed in a different ticket. In this ticket, the
issue is that the "other" field in Person was causing the
equals method of Person to always fail, and thus the
list of members being forced into a Set would fail because
two people who were identical were appearing not to be. The
backend would rightfully reject this request.

Fixes issue #382