Tests for discovery
This is around the investigation of issue #544. So far this adds tests
to mimic the issue in the javascript SDK, but these tests are passing indicating
that the issue probably doesn't lie with the repository. The next step is to look
at the node client itself in case it is causing some issue as the app-stack is confirmed
to be working properly.