[feature] Support Heartbeats as an alternative or alongside dropping connections
Dropping connections for SSE is best used for when SSE is being cached and for dealing with ghost
connections. If there is no cache layer, it is worthwhile to be able to remove the drop connections
and simply use a heartbeat to manage connections instead. This PR adds heartbeat capability to allow
both a heartbeat and dropping connections, and _only_ a heartbeat to manage connections for SSE.