[internal] fixing the initial setup and use of the Rocket quick-start (#740)
* [internal] fixing the initial setup and use of the Rocket quick-start

The Rocket Quick Start needs some care and attention.

* [internal] fixing the focus issue around the dialogs/rocket

The Rocket has a few issues with its state, but that state update isn't required
when the rocket isn't there. These changes update for the rocket when its
open. This also addresses most of the issue around the focus for dialogs.

* minor fixes

- cleanup of focus issue,
- make sure the "All Set" page is properly sized so it doesn't flicker around