[feature] Support Cache Control for website and APIs (#751)
This allows three different kinds of cache control to be set. Cache on the
GET requests for the APIs is force set to no-cache. Cache on the /index.html
is also set to no-cache. All other assets for the web that are delivered are
identified cacheable. All cache headers can be replaced by the user.