[bugfix/service-account-publish] Some dbs miss newly associated SA/Environments
While testing support for MySQL - it became clear that unless the relationship
between a service account and an environment is committed first, the subsequent
query will not pick up the change. Environment and Feature update publishing
operate by saving first and then async updating, but Service Accounts did not.
This fix ensures that the changes are in the database so that all databases
and their transactional modes will pick up the updating of a service account.