[feature/ha-support] Support for Postgres HA
High Availability does not function in the appropriate manner for FHOS
because of the way the REST layer interacts with the database layer. Transactions are short
and directed normally, but this causes problems with HA as each request can go to a different
backend which may not have the data yet.