[feature/auditing] provides transparent auditing and diffing
This allows features to be updated in part even when their versions are older. It is able
to check if there has been a clash between the version that is being updated and the current
version.

can be rolled out in stages where the historical versions are captured before the diffing starts.

todo:
- jersey needs to support deepEquals instead of just equals
- extensive "unit" testing needs to be enabled