[feature/enricher] Support enrichment pipeline
This change is designed to support the ability for an updated feature
to be picked up and enriched with all of the environment detail in a scalable
fashion. The feature ensures the local dacha is up to date and then adds
in the environment details and all of the existing features for that environment
yet keeps track of the features that were modified.
start of testing