[bugfix/dashboard-refresh] Fix refresh of dashboard (#959)
The shared state for which currently selected portfolio
and application were a bit messy and spread around. This had
led to some confusion when the app refreshes about which
portfolio and application it should be using. This tidies that
code up, centralises it, sorts out how it is initialized and
used and makes it work in a predictable way.

Fixes #954